Fixed various tests
- Regenerated Span model for track view testing
- Fixed "id" problems in SpanFitInformation
- Fixed wrong FQNs in SpanFitInformationTest

+++ src/main/kotlin/org/jetbrains/bio/statistics/gson/GSONUtil.kt
@@ -75,7 +75,7 @@ object GSONUtil {
                 val fqnElement = element.asJsonObject.remove(fqnField)
                 if (fqnElement == null) {
                     deserializationError("Class name (%s) field is missing." +
-                                         " Please, recalculate the model.", fqnField)
+                            " Please, recalculate the model.", fqnField)
                 }
                 val fqn = fqnElement!!.asString
 
@@ -99,17 +99,21 @@ object GSONUtil {
     }
 
     /**
-     * Same as [classAwareAdapter], but also saves the contents of a static field named "VERSION" and later
+     * Same as [classAwareAdapter], but also saves the contents of a static int field named "VERSION" and later
      * checks it against the actual contents of this field when loading.
      * Helps to work around the changes in the class field layout;
      * a version change is more self-explanatory than a random field reading failure.
      */
-    @JvmStatic fun <T : Any> classAndVersionAdapter(gson: Gson, factory: TypeAdapterFactory,
-                                                    fqnField: String, versionField: String): TypeAdapter<T> {
+    @JvmStatic fun <T : Any> classAndVersionAdapter(
+            gson: Gson,
+            factory: TypeAdapterFactory,
+            fqnField: String,
+            versionField: String
+    ): TypeAdapter<T> {
         return object : TypeAdapter<T>() {
             override fun write(out: JsonWriter, value: T) {
                 val modelFQN = value.javaClass.name
-                val modelVersion = getSerializationFormatVersion(value.javaClass)
+                val modelVersion = getSerializationFormatVersion(value.javaClass).toInt()
 
                 val delegate = gson.getDelegateAdapter(factory, TypeToken.get(value.javaClass))
 
@@ -148,7 +152,7 @@ object GSONUtil {
                 val recentVersion = getSerializationFormatVersion(aClass)
                 if (recentVersion != vers) {
                     deserializationError("Format has changed, '%s' expects '%s' version, but got '%s'",
-                                         fqn, recentVersion, vers)
+                        fqn, recentVersion, vers)
                 }
 
                 val adapter = gson.getDelegateAdapter(factory, TypeToken.get(aClass))
@@ -163,8 +167,8 @@ object GSONUtil {
                     return (aClass.getDeclaredField("VERSION").get(null)).toString()
                 } catch (e: Exception) {
                     deserializationError("Cannot get serialization format version." +
-                                         " Probably VERSION field is missing in %s. Exception message: %s",
-                                         aClass.name, e.message, cause = e)
+                            " Probably VERSION field is missing in %s. Exception message: %s",
+                        aClass.name, e.message, cause = e)
                     return "" // this statement can never be reached
                 }
             }
